Create template for experts page.
This commit moves the "Meet The Experts" content into its own template.
I did this because we decided to add videos to the listings of each
experts, and this was the easiest way to mark this up. This content
won't change frequently, so there isn't much change in terms of
usability.